write_aiger using CertifId and unsigned instead of uint64_t

+++ src/aig/aig/aig.h
@@ -30,7 +30,6 @@
 #include <stdlib.h>
 #include <string.h>
 #include <assert.h>
-#include <stdint.h> //@ needed for uint64_t : CertifId
 
 #include "misc/vec/vec.h"
 #include "misc/util/utilCex.h"
@@ -91,7 +90,7 @@ struct Aig_Obj_t_  // 8 words //@ no longer true because of CertifId
     };
 
     //@ A permanent id used by certificates to refer to this node while the AIG is evolving.
-    uint64_t         CertifId;
+    unsigned         CertifId;
 };
 
 // the AIG manager
@@ -288,7 +287,7 @@ static inline int          Aig_ObjIsChoice( Aig_Man_t * p, Aig_Obj_t * pObj )   
 static inline int          Aig_ObjIsCand( Aig_Obj_t * pObj )      { return pObj->Type == AIG_OBJ_CI || pObj->Type == AIG_OBJ_AND || pObj->Type == AIG_OBJ_EXOR;     }
 static inline int          Aig_ObjCioId( Aig_Obj_t * pObj )       { assert( !Aig_ObjIsNode(pObj) ); return pObj->CioId;                                            }
 static inline int          Aig_ObjId( Aig_Obj_t * pObj )          { return pObj->Id;                     }
-static inline uint64_t     Aig_ObjCertifId( Aig_Obj_t * pObj )    { return pObj->CertifId;               } //@ getter
+static inline unsigned     Aig_ObjCertifId( Aig_Obj_t * pObj )    { return pObj->CertifId;               } //@ getter
 
 static inline int          Aig_ObjIsMarkA( Aig_Obj_t * pObj )     { return pObj->fMarkA;  }
 static inline void         Aig_ObjSetMarkA( Aig_Obj_t * pObj )    { pObj->fMarkA = 1;     }
